requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
streamlit>=1.28.0
numpy>=1.24.0
Pillow>=10.0.0
flask>=3.0.0
tensorflow>=2.15.0